[ARTS-423] changes to ensure json is piked up, rename of package, removing properties from service
