+++ pom.xml
@@ -35,7 +35,7 @@
         <module>sample-service</module>
         <module>practitioner-service</module>
         <module>trial-config-service</module>
-        <module>sc-config-server</module>
+        <module>mts-config-server</module>
     </modules>
 
     <dependencies>
