--- sc-config-server/src/main/java/uk/ac/ox/ndph/mts/sc_config_server/package-info.java
@@ -1,4 +1,0 @@
-/**
- *
- */
-package uk.ac.ox.ndph.mts.sc_config_server;
