--- sc-config-server/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-#spring.application.name=spring-config-server
-server.port=8888
-
-spring.cloud.config.server.git.uri=https://github.com/NDPH-ARTS/spring-config-server.git
-spring.cloud.config.server.git.default-label=main
-#spring.cloud.config.server.git.clone-on-start=true
-
-#spring.cloud.config.server.git.search-paths=trial-config*
-#spring.cloud.config.server.git.ignore-local-ssh-settings=false
-#spring.cloud.config.server.git.strict-host-key-checking=false
-#health.config.enabled=false
-#spring.cloud.config.server.git.skip-ssl-validation=true
