+++ trial-config-service/src/main/java/uk/ac/ox/ndph/mts/trial_config_service/controller/TrialConfigController.java
@@ -1,9 +1,11 @@
 package uk.ac.ox.ndph.mts.trial_config_service.controller;
 
-
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cloud.context.config.annotation.RefreshScope;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,95 +14,89 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientException;
-import reactor.core.publisher.Mono;
-import uk.ac.ox.ndph.mts.trial_config_service.config.GitRepo;
 import uk.ac.ox.ndph.mts.trial_config_service.exception.InvalidConfigException;
 import uk.ac.ox.ndph.mts.trial_config_service.exception.ResourceAlreadyExistsException;
 import uk.ac.ox.ndph.mts.trial_config_service.model.Trial;
 import uk.ac.ox.ndph.mts.trial_config_service.service.TrialConfigService;
+import java.io.File;
 
-import java.io.IOException;
-
-
-
+@RefreshScope
 @RestController
 @RequestMapping("/trial-config")
 public class TrialConfigController {
 
+    @Value("${spring.cloud.config.uri}")
+    private String configURI;
+
+    @Value("${spring.application.name}")
+    private String applicationName;
+
+    @Value("${spring.cloud.config.profile}")
+    private String profile;
+
+    @Value("${spring.cloud.config.label}")
+    private String label;
+
     private final TrialConfigService trialConfigService;
     private final WebClient webClient;
-    private GitRepo gitRepo;
 
     //TODO(darrensmithson) - only temp until we decide where to get userId from
     String userId = "admin001";
 
     @Autowired
-    public TrialConfigController(TrialConfigService trialConfigService, WebClient webClient, GitRepo gitRepo) {
+    public TrialConfigController(TrialConfigService trialConfigService, WebClient webClient) {
         this.trialConfigService = trialConfigService;
         this.webClient = webClient;
-        this.gitRepo = gitRepo;
-    }
-
-    public TrialConfigController(TrialConfigService trialConfigService, String baseUrl) {
-        this.trialConfigService = trialConfigService;
-        this.webClient = WebClient.create(baseUrl);
     }
 
     //TODO(darrensmithson) - we are using a GET and POST for similar functionlaity to accomodate
     // early testing needs, once we are closer to final developmemnt we should ensure only 1 POST is used.
     @GetMapping("/trial")
-    public Trial saveTrialFromGit(
+    public Trial saveTrialFromConfigServer(
             @RequestParam
-                    String filename) throws InvalidConfigException, ResourceAlreadyExistsException {
-        return trialConfigService.saveTrial(createTrialFromGitRepo(filename), userId);
+                    String filename)
+        throws InvalidConfigException, ResourceAlreadyExistsException {
+        return trialConfigService.saveTrial(createTrial(filename, repoURL()), userId);
     }
 
     @PostMapping("/trial")
     public Trial saveTrialFromJson(
             @RequestBody
                     String jsonData) throws InvalidConfigException, ResourceAlreadyExistsException {
-        return trialConfigService.saveTrial(createTrialFromJsonData(jsonData), userId);
+        return trialConfigService.saveTrial(createTrial(jsonData), userId);
     }
 
-    protected Trial createTrialFromGitRepo(String fileName) {
-        Trial trial;
-
-        try {
-            byte[] fileBytes = gitRepo.getTrialFile(fileName);
-            ObjectMapper objMapper = new ObjectMapper();
-            trial = objMapper.readValue(fileBytes, Trial.class);
-        } catch (IOException ioException) {
-            throw new InvalidConfigException(ioException.getMessage());
-        }
-
-        return trial;
-    }
-
-    protected Trial createTrialFromJsonData(String trialConfig) {
+    protected Trial createTrial(String jsonData) {
         Trial trial;
 
         try {
             ObjectMapper objMapper = new ObjectMapper();
-            trial = objMapper.readValue(trialConfig, Trial.class);
+            trial = objMapper.readValue(jsonData, Trial.class);
         } catch (JsonProcessingException jpeEx) {
             throw new InvalidConfigException(jpeEx.getMessage());
         }
         return trial;
     }
 
-    protected Mono<Trial> createTrialFromURL(String trialConfig) {
-        Mono<Trial> response;
+    protected Trial createTrial(String trialConfig, String url) {
+        Trial trial;
 
         try {
-            response = webClient.get()
-                .uri(trialConfig)
+            trial = webClient.get()
+                .uri(url + File.separator + trialConfig)
+                .accept(MediaType.APPLICATION_JSON)
                 .retrieve()
-                .bodyToMono(Trial.class);
+                .bodyToMono(Trial.class)
+                .block();
         } catch (WebClientException wceEx) {
             throw new InvalidConfigException(wceEx.getMessage());
         }
 
-        return response;
+        return trial;
+    }
+
+    private String repoURL() {
+        return configURI + File.separator + applicationName + File.separator + profile + File.separator + label;
     }
 
 }
